+++ packages/pangea-sdk/PangeaCyber.Net/Response.cs
@@ -1,6 +1,6 @@
-using System.Net.Http;
 using Newtonsoft.Json;
 using PangeaCyber.Net.Exceptions;
+using System.Net.Http;
 
 namespace PangeaCyber.Net
 {
@@ -19,6 +19,9 @@ namespace PangeaCyber.Net
 
         ///
         public HttpResponseMessage HttpResponse { get; set; } = default!;
+
+        ///
+        public List<AttachedFile> AttachedFiles = new();
 
         ///
         public Response()