+++ packages/pangea-sdk/CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Documentation for client builders and constructors.
+- Audit /download_results endpoint support
 
 ### Changed