+++ packages/pangea-sdk/CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Documentation for client builders and constructors.
 - Audit /download_results endpoint support
 
+### Fixed
+
+- Put to presigned url. It should just put file in raw, not in form format.
+
 ### Changed
 
 - Skip sending `filename` field when posting to presigned URL. Make GCP S3 to return a `BadRequest`